Move languageDescriptorMap
